test(nat): Add tests for ICMP Error messages (translation, checksum)
Add unit tests and Bolero tests for checksum update and validation for
ICMP Error messages. The checks relate to the checksums for the outer
IPv4 header (if any), the ICMP header, and the inner IPv4 header (if
any), and for translating addresses and port in the inner packet.

More generally, these tests also ensure that we parse and deparse
ICMP Error messages' inner packets correctly.

Signed-off-by: Quentin Monnet <qmo@qmon.net>