fix(net): Fix ICMP Error messages checksum payload computation
Some headers such as TCP, UDP, ICMP, have their checksum computed based
on their payload. In the case of ICMPv4 and ICMPv6, this payload may
contain an inner IP packet (possibly truncated).

Now that we parse this inner IP packet, we've got a problem: we store
the inner headers in dedicated structures, and the "payload" for the
packet is the payload _under these headers_, meaning that it's no longer
equal to the outer transport payload - in other words, packet.payload()
is no longer equal to the ICMP header's payload.

As a consequence, we need to regenerate a payload for ICMP Error
messages by de-parsing the embedded headers to a buffer, followed by the
packet's payload. This is painful, but not really avoidable.

Note: In the future we might want to refactor this step as part as the
Icmp4Checksum and Icmp6Checksum generation; but these objects take
pointers to existing payload (they don't store a vector of their own)
and this would involve too much refactoring, so we leave it out for now.

Also drop IcmpAnyMut's implementation of is_error_message(), which is
unused at the moment. We can always re-add it later if necessary, it's
identical to IcmpAny's implementation anyway.

Fixes: 85f935e9027b ("feat(net): Update embedded IP's checksum in Headers' update_checksum()")
Signed-off-by: Quentin Monnet <qmo@qmon.net>