fix(net): Fix deparsing for ICMP Error messages' inner headers
When we introduced inner headers deparsing for ICMP Error messages, we
naturally added the deparsing step at the end of the DeParse
implementation for Headers. This was a mistake! If the Headers have no
UDP encapsulation header, then the function returns, and we never
process the embedded headers, resulting in an imcomplete packet.

Instead, do not return early if there's no UDP encapsulation header, and
keep going until we've processed ICMP inner headers, too. Add sanity
checks about parent headers.

Fixes: 684c65ca41f6 ("feat(net): Introduce EmbeddedHeaders to parse ICMP's inner packets")
Signed-off-by: Quentin Monnet <qmo@qmon.net>