fix(net): Fix checksums update order for ICMP Error msg's inner headers
When we added support for ICMP Error messages' inner headers checksum
update, we added the udpate at the end of the function to update
headers' checksums, below the update of the transport-layer header's
checksum. This sounds logical: we process each layer in order. But it is
wrong: if we update the checksums of the inner layers after the one for
the outer transport layer, we have an issue. This is because the inner
headers are part of the ICMP payload, and therefore, they need to be
accounted for when computing the checksum for the ICMP header. If we
update these headers _after_ updating the ICMP checksum, we invalidate
this checksum.

Reverse the order to make sure that we update the inner headers before
we update the ICMP checksum.

Fixes: 85f935e9027b ("feat(net): Update embedded IP's checksum in Headers' update_checksum()")
Signed-off-by: Quentin Monnet <qmo@qmon.net>