feat(nat): Pass flow timeout via flow session value, use on ts update
We recently introduced a way for users to specify a duration for flows
idle timeout when configuring stateful NAT. We need to actually use the
provided values.

As timeouts are defined per VpcExpose, we need to keep track of what
idle timeout value a given flow should use. One way to do this is to
pass the expected timeout value as part of the FlowInfo object (the
value from session table lookups), so that we can immediately use it on
successful lookups to update the remaining time of life for the flow.

Note that because of our current design, we have a single flow entry for
both source and destination NAT, with the "rule" for source NAT coming
from the VpcExpose for the VPC the packet is leaving, and the one for
destination NAT based on the VPC the packet is heading to. Of course we
can't apply both values if they differ: in this case, we chose to apply
the smallest one. This design will likely evolve in the future, as we
plan to forbid setting up stateful NAT in both directions at the same
time.

Signed-off-by: Quentin Monnet <qmo@qmon.net>