refactor(nat): Refactor NAT pools building
It's been bothering me to have so much near-duplication of the code for
building NAT pools for source and destination NAT, and for NAT44 and
NAT66 (four nearly-identical snippets). Let's refactor them using a
generic function. The signature of this generic function is not trivial
(Claude helped). Nevertheless, the results is easier to read, as it's
much easier to spot the differences between the different IP version and
direction cases.

Signed-off-by: Quentin Monnet <qmo@qmon.net>