test(nat): Add test for unidirectional NAT
Check what happens with stateful NAT when we have rules for translation
in only one direction. When this direction is source NAT, we should
translate the source address, and forward the packet, as expected. When
this direction is destination NAT only, we should reject the connection:
the packet should remain unchanged, and be marked as "done" (meaning a
follow-up stage should drop it).

Signed-off-by: Quentin Monnet <qmo@qmon.net>