feat(nat): Add logic for processing ICMP in stateful NAT
We've designed the stateful NAT's allocator with TCP and UDP in mind,
with two ports. This commit adds support for translating ICMP, too.

Ideally, we should have a dedicated structure adapted for ICMP, reusing
some of the components of the TCP/UDP structures. We're not doing that:
we simply reuse the existing allocator, using just one of the two fields
to hold the identifier. It is expected that this code will be refactored
in the near future.

Signed-off-by: Quentin Monnet <qmo@qmon.net>