feat(nat): Allow null identifiers in stateful NAT's allocators
We've designed the stateful NAT's allocator with TCP and UDP in mind,
with port values that cannot be zero. Now that we're working with ICMP
as well, we may have ICMP identifiers that are zero. We need to support
these, in particular:

- We cannot mark value 0 as reserved in the port allocator for ICMP,
  otherwise we will not be able to reserve the corresponding port for
  the return session if a packet arrives with a null identifier.

- We cannot consider the value as NonZero when we allocate it, because,
  well, now it can be zero.

We still want the constraint to be enforced for TCP and UDP, though. To
accommodate for the different transport protocols, we pass a boolean
flag via the arguments of the different allocation functions. This is
not elegant, but it works. It is expected that the allocator will be
reworked in the near future, anyway.

Signed-off-by: Quentin Monnet <qmo@qmon.net>