fix(net): ICMP-in-ICMP for NAT: try harder
When we added support for parsing inner ICMP packets inside of ICMP
Error messages, we missed a number of additional changes that also need
to happen in the code, in particular if we want headers parsing to work
as expected. Discovered while wondering why Bolero tests wouldn't
produce inner ICMP headers, ever.

Fixes: f9652c527b2e ("feat(net): Support ICMP-in-ICMP")
Signed-off-by: Quentin Monnet <qmo@qmon.net>