bump(n-vm): to v0.0.9 to avoid test breakage
This fixes a recent (and poorly understood) test regression which is blocking CI.

n-vm is still very young, but it is essential for dpdk testing. It isn't ready for wide scale use in this code base
outside of development settings, but the newest version gets it closer.

(n-vm is still used constantly in development flows)

Signed-off-by: Daniel Noland <daniel@githedgehog.com>