feat(nat): Add debug logs on session creation
It takes a whole commit, because we need some pretty-printing in order
to get single-line printing for new session keys and values.

The implementation of the Display traits is not always 100% complete
(regarding all data available in the structs), but it should cover our
immediate needs, and we can refine it later if necessary.

Signed-off-by: Quentin Monnet <qmo@qmon.net>