fix(routing,dataplane): concurrent use fibreader
When interfaces are attached to a VRF, interface objects are
equipped with a fibreader for the fib correspoding to a VRF. This
is not correct since FibReaders should not be shared by distinct
threads. Fix by storing, instead, the key of the fib.

The original idea of storing a fibreader for an interface was that,
once the interface was looked up, no further lookups would be
needed to access a particular fib. That advantage is no longer
such due to the fact that the logic is partitioned in distinct
stages. So, storing a fibkey is much simpler and, with the local
cache of fib-readers, the extra lookup that would have been saved
is possibly very cheap.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>