test(net,nat): Support inner ICMP in ICMP Error messages generation tool
In net's test utilities, function
build_test_icmp4_destination_unreachable_packet() produces an ICMP Error
message packet with an embedded TCP packet. Extend the function so that
the caller can make it use a TCP, UDP, or ICMPv4 inner packet instead.

The test in stateless NAT is adjusted for compatibility, but keeps using
a TCP inner packet.

This is in prevision for more tests, with inner ICMP packets for
example.

Signed-off-by: Quentin Monnet <qmo@qmon.net>